% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rainresults.R
\docType{data}
\name{rainresults}
\alias{rainresults}
\title{Memory parameter estimates and diagnostics for precipitation time series}
\format{
A dataset with 2118 rows and 19 variables containing:
\describe{
\item{statID}{station identifier}
\item{lat}{latitude (CRS WGS84)}
\item{lon}{longitude (CRS WGS84)}
\item{alt}{altitude (in meters)}
\item{koppen}{Koppen climate zone classification}
\item{zone_c}{Koppen main climate zone}
\item{zone_p}{Koppen precipitation subgroup}
\item{zone_t}{Koppen temperature subgroup}
\item{quspurious5}{indicator variable if null of \insertCite{qu2011test}{prcpFC} test is rejected on 5\% significance level}
\item{quspurious10}{indicator variable if null of \insertCite{qu2011test}{prcpFC} test is rejected on 10\% significance level}
\item{meanbreak}{indicator variable if null of \insertCite{shao2011simple}{prcpFC} and \insertCite{iacone2014fixed}{prcpFC} tests is rejected on 5\% significance level}
\item{armanoise}{}
\item{potentially_contaminated}{indicator variable if any of the employed tests rejects its null hypothesis on 10\% significance level}
\item{LW}{memory parameter estimate according to the Local Whittle estimator of \insertCite{kunsch1987statistical}{prcpFC}}
\item{LWmixed}{memory parameter estimate according to the Local Whittle estimator of \insertCite{kunsch1987statistical}{prcpFC} adjusted for short-run dynamics}
\item{houLW}{memory parameter estimate according to the modified Local Whittle estimator of \insertCite{hou2014modified}{prcpFC}}
\item{houLWmixed}{memory parameter estimate according to the modified Local Whittle estimator of \insertCite{hou2014modified}{prcpFC} adjusted for short-run dynamics}
\item{bestd}{memory parameter estimate based on the diagnostics}
\item{ddiff}{difference of original and modified memory parameter estimate}
#' }
}
\usage{
data(rainresults)
}
\description{
Dataset with memory parameter estimates and diagnostics on
the underlying time series for precipitation data collected
by the monitoring stations of the Global Historical Climate
Network daily (GHCNd). The memory parameter estimates are
based on data from January 1960 until December 2019. The
dataset contains the variables station identifier, latitude,
longitude, altitude, Koppen climate zone classification,
Koppen main climate zone, Koppen precipitation subgroup,
Koppen temperature subgroup, indicator variables based on
the methodology described in \insertCite{fritsch2025}{prcpFC}
for spurious long memory, mean breaks, short memory noise,
and potential contaminations, different sets of memory
parameter estimates and differences in the estimates.
}
\details{
For more details on data collection by the GHCNd, see
\insertCite{menne2012overview}{prcpFC},
\insertCite{durre2010comprehensive}{prcpFC}, and
\insertCite{durre2008strategies}{prcpFC}.
The dataset is employed in \insertCite{fritsch2025}{prcpFC}
to analyze the memory properties of the time series and
generate forecasts based on the memory parameter estimates.
The climate zone classification assigned to the stations
was derived by \insertCite{kottek2006}{prcpFC}.
}
\examples{
\dontrun{
  data(rainresults, package = "prcpFC")
  n <- rainresults$bestd
  median(n)
}

}
\references{
\insertAllCited{}
}
\keyword{datasets}
